% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcausality-package.R
\docType{package}
\name{ForCausality}
\alias{ForCausality}
\alias{ForCausality-package}
\title{ForCausality: A Curated Collection of Causal Inference Datasets and Tools}
\description{
{
Provides a comprehensive set of datasets and tools for causal inference research. 
The package includes data from clinical trials, cancer studies, epidemiological surveys, environmental exposures, and health-related observational studies.
}
}
\details{
ForCausality: A Curated Collection of Causal Inference Datasets and Tools

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Curated Collection of Causal Inference Datasets and Tools
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/Toby-codigos/ForCausality}
}
}
}
\author{
{
\strong{Maintainer}: Tomás Valderrama \email{tomasvm2004@gmail.com}
}
}
